requirements/test.txt

# includes the base.txt requirements file
-r base.txt

# pytest==7.1.3
pytest-django==4.5.2

